wrote a new function to get a user id for a given username
I have wrote a new function to find a user id for a given username (this is in the main.go file in the utils folder).

In my addFriend.go file in the actions folder, I use this new function that I wrote to put the data for the friend request into the friend request table.