I now show a message to the user and if the user answers the question with yes or no a particular event will occur
I now show the user a message when they have sent the friend request.

I made a function to ask the user do they want to accept or decline the request.

if the status is pending the user gets ask the question from the function that I wrote, if they say something like yes, the status is changed and if they say something like no, the status is also changed.

I need to write some new functions to update the table if the answer is yes, if it is no I need to delete that user from the table.