got add a friend and send friend request working with html and I also did some tidying up.
I have finally got the html working for add a friend and accept or decline friend request for the user (this took hours), if you enter "accept" certain queries will execute and this will also happen if you enter "decline".

I have added in some new bits of html to make it look nicer for the user and I have also removed anything that I no longer needed such as (code and comments).