wrote 5 new functions.
today (19/01/25) I have wrote 5 new functions.

InsertLoggedInUserToTable. this function puts whatever user has logged in to a new table called loggedin.

RetrieveEmail. this function gets a username depending on what the email is.

RetrieveUsername. this function gets an email depending on the username is.

GetLastUserLoggedIn. this function gets the last logged in user from the new table called loggedin.

RetrieveEmailFromId. this function gets an email depending on what id value is given to it.