updated UI variable and wrote function to get friends
I have now added in more html into the UI variable, this is then used in my login code.

In my main.go file I have wrote a function to get all of your friends, but it was showing duplicates, I found this code online that was quite useful for removing duplicates.