removed wrong html and wrote two new functions
I have removed the html for the user to enter either accept or decline I removed this because this is not correct. any of my code that I was using is now commented out (if I need it later on)
.
In my main.go file I have wrote two new functions one of the functions gets toUserName from the database depending on what you pass as the username and the other function gets the status for the logged in user, toUserName where the status is pending.