wrote new function to get value of what user enters, can get all pending requests for the logged in user and deleted some code and comments that were no longer needed.
I wrote a new function to get the value of what the user enters in if they want to accept or decline the request, I did this because it was NOT letting me get the value out of it in a variable with html content.

When I call my function called GetPendingRequestsForLoggedInUser this function gets all the pending friend requests for the logged in user.

I also removed some code that I no longer needed and I also did this for some of the comments I did have.