wrote two new functions and wrote code to create a channel for a certain severId.
In my main.go file I have wrote two new functions:
GetServerId: this function gets the last serverId that was made.

InsertChannelName: this function inserts the name of a channel to a certain serverId.

In my createChannel.go file I extract the name of the channel and I get the last serverId using this function: GetServerId then I insert this data to my channel table in my database using this query: InsertChannelName.