wrote a new function to add friend to a server
In my utils/main.go file I have wrote a new function to add a friend to a server.

In my addFriendToServer.go file I am pulling down the last logged in users "friends" and "severs". When a friend name and server name are clicked they are inserted into the function that I wrote, in my main.go file, this function is called: AddFriendToServer.