wrote a new function and did some refactoring.
In my main.go file I have wrote a new function to get all of the friendnames that are in a server.

In my addFriendToServer.go file I have refactored my code to now be able to insert many people into a server, not limiting it to only one friend like it was previously.