only the creator of the server can now add friends to the server and remove friends from the server.
In my main.go file I have wrote a new function to get all of the names of people who have created a server.

In my ui.go file I have created a new variable to show that you are the owner of the server.

In my login.go file I have wrote some new code to show if you are the owner of a server, if you are not the owner of the server you are shown nothing because you have not created a server.