wrote 5 more functions, with a brief explanation on what each function is doing.
In my main.go file I have wrote the following functions:

GetOwner: in this function I get the owner id of the last server that was created.

GetOwnerNameFromId: in this function I pass the id from the previous function into this function, this then gives me the owner of the server.

GetSenderOfChannelMessage: in this function I retrieve the sender of who sent the message to a channel.

ShowMessageInChannel: in this function I show the messages sent from whoever is in the array.

Run: in this function I use all of the above functions together to get a result back.

In my login.go file I call the Run method to see the results.