did some refactoring, wrote 2 new functions, then these functions are then used in my login.go file.
In my main.go file I have removed the "," and replaced it with a ":".

I wrote a new function to get all of the friends that have been added to a server.

I wrote another function to pull all of the messages in a channel within a server and only show them to the owner of the server or if they have been added into a sever.

In my login.go file I have wrote some code to only show the owner of the server and however has been added into the server, the messages within a channel in a server.