combining go files together.
I wrote a new function to redirect a user to a certain go file that is running.

I have two variables called addF and removeF these are then added into an array then I can chose what page the user is redirected to, with the use of the index in the array.