created two new variables, wrote four new functions and now have enforced strict usernames
In my ui/ui.go file I have created two new variables, these are used to show to the user.

In my main.go file I have wrote 4 new functions. below is a small description on what each of them do:

GetChannelsInServer: this function gets all of the channels that are in a server.

GetMessagesInChannel: this function gets the messages that are in a channel.

GetServerIdFromServerName: this function gets me a server id from a server name.

GetCommunicatorsUsernames: this function gets all of my users that have created an account.

In my makeAccount.go file I have wrote some code to enforce strict usernames, meaning if you enter a name and it is in my db you cannot use that name, you need to use a different name.