working on ui and styling
In my ui.go file I have added in a bit of css to make the ui look much nicer for the user.

In my makeAccount.go file I have added in a good bit of css and I have also added in some icons to make it pop a bit more. I also added in a new line of code to allow the user to go the login page after they have created there account.

I have created a new go file called "welcomeToKC.go", in this file I just show the user a welcome page and on this page they can chose to either make an account or login into there existing account.