working on ui & styling
In my addFriend.go file I have added in a good bit of css to make the ui better.

In my createServer.go file I have added in a good bit of css to make the ui better, I have also added in a confirmation message to the user, telling them that there server is created.

In my sendMessage.go file I have added in a good bit of css to make the ui better, I have also added in a confirmation message to the user, telling them that there message has been sent to there friend.