working on ui, debugging and removing code that is no longer needed
I have added in a good bit of css to make the ui nicer for the user.

I have spent the last 2 hours trying to get the h1 tag at the top of the page but it is not working at the moment.

I have also removed the two friends sections, there is only one friends section now (like it should be)!