adding in edge cases for add friend (send friend request) and working on ui and styling
In my addFriend.go file I have added in some code to stop my system from breaking when you enter in a name that is not in my db.

I have also added in some more code, to not allow you to send a friend request if you are friends with who you are trying to send the friend request to.

In my ui.go file I have added in some css to make the ui better for the user.