fixed docker issues with timestamp and working on styling still and I have created buttons for accepting or declining a friend request
In my main.go file I have did a refactor, I had to do this because when I was running my code via a docker container, it was not showing the timestamp in the correct way.

In my login.go file I have removed the need for the user to type in accept or decline for the friend request, they now just need to click on the accept or decline button.

I am also working on the styling still for Kyle Connect.