small refactor, added in more css and changed the timestamp location and I have also created a docker-compose file.
In my main.go file I have did a small refactor.

In my sendMessageToChannelInServer.go file I have added in more css and I have changed the timestamp to be Ireland time, previously it was not Ireland time.

In my sendMessage.go file I have changed the timestamp to be Ireland time, previously it was not Ireland time.

I have created a docker-compose file, this file is used to spin up all of my docker containers.